client: ThreadHeader is its own component
