+++ client/src/components/ReviewThreads/Thread/Thread.css.js
@@ -7,24 +7,6 @@ const useStyles = makeStyles((theme) => ({
     position: "relative",
   },
 
-  header: {
-    padding: "40px",
-    borderBottom: `1px solid ${theme.palette.secondary.lightGray}`,
-    // display: "flex",
-    // justifyContent: "space-between",
-    // alignItems: "center",
-  },
-
-  headerActionButton: {
-    margin: "0 4px",
-  },
-
-  title: {
-    margin: 0,
-    fontWeight: "bold",
-    fontSize: "20px",
-  },
-
   syntaxWrapper: {
     padding: "3rem",
   },
