+++ client/src/components/ReviewThreads/ThreadHeader/index.js
@@ -1,0 +1,3 @@
+import ThreadHeader from './ThreadHeader';
+
+export default ThreadHeader;
